pyproject.toml
[tool.poetry]
name = "marketplace-backend"
version = "0.1.0"
description = ""
authors = ["Andrey Latypov <andrey.nasayer232@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
djangorestframework = "^3.13.1"
django-environ = "^0.9.0"
drf-spectacular = "^0.23.1"
django-db-logger = "^0.1.12"
psycopg2-binary = "^2.9.3"
djangorestframework-simplejwt = ">=4.3.0"
drf-nested-routers = "^0.93.4"
djoser = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^5.0.4"
black = "^22.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
force-exclude='/migrations/'

[tool.isort]
profile="black"
filter_files="true"
known_django = "django"
known_first_party = 'apps'
sections = ['STDLIB', 'DJANGO', 'THIRDPARTY', 'FIRSTPARTY', 'LOCALFOLDER' ]
length_sort="True"
skip_glob=["*/migrations/*"]
